fix(auth): sync Privy sign-out across the whole app, not just /login
The Privy <-> wallet-store bridge only lived inside PrivyLogin, which is
unmounted everywhere except /login. Signing out from a dashboard page left
Privy logged out internally while the wallet store still reported
connected, so the UI never reflected the sign-out.

Moved the bridge into its own component (PrivyWalletBridge) mounted once
at the app root, so it keeps the wallet store in sync regardless of route.
PrivyLogin is now just the sign-in/sign-out button.
